upgrade faiss #7

+++ src/index/mod.rs
@@ -171,7 +171,7 @@ pub trait Index {
     fn reset(&mut self) -> Result<()>;
 
     /// Remove data vectors represented by IDs.
-    fn remove_ids(&mut self, sel: &IdSelector) -> Result<i64>;
+    fn remove_ids(&mut self, sel: &IdSelector) -> Result<usize>;
 }
 
 /// Sub-trait for native implementations of a Faiss index.
